+++ data/team.js
@@ -1,0 +1,153 @@
+import avatar1 from '../src/assets/images/team/member1.png';
+import avatar2 from '../src/assets/images/team/member2.png';
+import avatar3 from '../src/assets/images/team/member3.png';
+import avatar4 from '../src/assets/images/team/member4.png';
+
+const data = [
+    {
+      id: 1,
+      avatar: avatar1,
+      name: 'Emmalee Mclain',
+      designation: 'Product Designer',
+      socialLinks: [
+        {
+          name: 'twitter',
+          link: 'http://twitter.com',
+        },
+        {
+          name: 'github',
+          link: 'http://github.com',
+        },
+        {
+          name: 'dribbble',
+          link: 'http://dribbble.com',
+        },
+      ],
+    },
+    {
+      id: 2,
+      avatar: avatar2,
+      name: 'Daisy Morgan',
+      designation: 'Lead developer',
+      socialLinks: [
+        {
+          name: 'twitter',
+          link: 'http://twitter.com',
+        },
+        {
+          name: 'dribbble',
+          link: 'http://dribbble.com',
+        },
+      ],
+    },
+    {
+      id: 3,
+      avatar: avatar3,
+      name: 'Ariyanna Hicks',
+      designation: 'Experience Designer',
+      socialLinks: [
+        {
+          name: 'twitter',
+          link: 'http://twitter.com',
+        },
+        {
+          name: 'github',
+          link: 'http://github.com',
+        },
+      ],
+    },
+    {
+      id: 4,
+      avatar: avatar4,
+      name: 'Yamilet Hooker',
+      designation: 'User interface designer',
+      socialLinks: [
+        {
+          name: 'twitter',
+          link: 'http://twitter.com',
+        },
+        {
+          name: 'github',
+          link: 'http://github.com',
+        },
+        {
+          name: 'dribbble',
+          link: 'http://dribbble.com',
+        },
+      ],
+    },
+    {
+      id: 5,
+      avatar: avatar1,
+      name: 'Emmalee Mclain',
+      designation: 'Product Designer',
+      socialLinks: [
+        {
+          name: 'twitter',
+          link: 'http://twitter.com',
+        },
+        {
+          name: 'github',
+          link: 'http://github.com',
+        },
+        {
+          name: 'dribbble',
+          link: 'http://dribbble.com',
+        },
+      ],
+    },
+    {
+      id: 6,
+      avatar: avatar2,
+      name: 'Daisy Morgan',
+      designation: 'Lead developer',
+      socialLinks: [
+        {
+          name: 'twitter',
+          link: 'http://twitter.com',
+        },
+        {
+          name: 'dribbble',
+          link: 'http://dribbble.com',
+        },
+      ],
+    },
+    {
+      id: 7,
+      avatar: avatar3,
+      name: 'Ariyanna Hicks',
+      designation: 'Experience Designer',
+      socialLinks: [
+        {
+          name: 'twitter',
+          link: 'http://twitter.com',
+        },
+        {
+          name: 'github',
+          link: 'http://github.com',
+        },
+      ],
+    },
+    {
+      id: 8,
+      avatar: avatar4,
+      name: 'Yamilet Hooker',
+      designation: 'User interface designer',
+      socialLinks: [
+        {
+          name: 'twitter',
+          link: 'http://twitter.com',
+        },
+        {
+          name: 'github',
+          link: 'http://github.com',
+        },
+        {
+          name: 'dribbble',
+          link: 'http://dribbble.com',
+        },
+      ],
+    },
+  ];
+
+  export default data;
